copied events now brag on slack too
